test(p2p): move sync state tests into the tests tree
