+++ crates/p2p/src/sync/broadcaster.rs
@@ -167,33 +167,3 @@ impl<T: P2PTransport> Broadcaster<T> {
         &self.transport
     }
 }
-
-#[cfg(all(test, feature = "libp2p-transport"))]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_create_broadcast() {
-        use std::str::FromStr;
-        let cid =
-            Cid::from_str("bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi").unwrap();
-        let block = b"test block data";
-        let doc_id = "bae-123";
-        let collection_id = "users";
-        let creator = "12D3KooWPeer";
-
-        let broadcast = Broadcaster::<crate::host::Libp2pTransport>::create_broadcast(
-            &cid,
-            block,
-            doc_id,
-            collection_id,
-            creator,
-        );
-
-        assert_eq!(broadcast.doc_id, doc_id);
-        assert_eq!(broadcast.collection_id, collection_id);
-        assert_eq!(broadcast.creator, creator);
-        assert_eq!(broadcast.block, block.to_vec());
-        assert_eq!(broadcast.cid, cid.to_bytes());
-    }
-}
