+++ crates/p2p/src/sync/peer_state/tracker/mod.rs
@@ -7,8 +7,6 @@
 
 mod memory;
 mod operations;
-#[cfg(test)]
-pub mod tests;
 
 use std::collections::{HashMap, HashSet, VecDeque};
 use std::time::{Duration, Instant};
