+++ crates/p2p/tests/sync/main.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "libp2p-transport")]
+mod broadcaster;
+mod peer_state;
